Revert "bufferSize and numOfBuffers (BufferManager) uint32_t -> size_t"
This reverts commit bee4b61e93bf0372631a08500fcc0fe136bbeb4a.

+++ nes-memory/include/Runtime/BufferManager.hpp
@@ -202,8 +202,8 @@ private:
 
     mutable std::recursive_mutex unpooledBuffersMutex;
 
-    size_t bufferSize;
-    size_t numOfBuffers;
+    uint32_t bufferSize;
+    uint32_t numOfBuffers;
 
     uint8_t* basePointer{nullptr};
     size_t allocatedAreaSize;
